connections work

+++ src/CircuitView/CircuitView.ts
@@ -107,7 +107,12 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
         this.blockArea = new BlockArea(this)
         this.inputArea = new IOArea(this, 'inputArea')
         this.outputArea = new IOArea(this, 'outputArea')
+
+        window.onkeydown = this.onKeyDown.bind(this)
+        window.onkeyup = this.onKeyUp.bind(this)
     }
+
+
     circuit: Circuit
     gridMap = new CircuitGrid()
     traceLayer: ICircuitTraceLayer
@@ -118,11 +123,13 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
 
     draggingMode: DraggingMode
     scrollStartPos: Vec2
-    connectingTrace: ID = -1
+    connectingTraceID: ID = -1
 
     selectedElements = new Set<CircuitElement>()
     selectedElementsInitPos = new Map<CircuitElement, Vec2>()
+    
     selectionBox: HTMLDivElement
+    selectionBoxInitPos: Vec2
 
     blocks = new Map<ID, FunctionBlockView>()
     traces = new Map<ID, CircuitTrace>()
@@ -131,7 +138,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
     loadCircuit(circuit: Circuit) {
         this.circuit = circuit
         this.circuit.onModificationUploaded = (type, success, blockID, ioNum) => {
-            console.log('Modification result:', {type, success, blockID, ioNum})
+            
         }
         console.log('CircuitView: Load circuit')
         const margin = vec2(6, 2)
@@ -175,18 +182,23 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
         this.circuit.connectFunctionBlockInput(inputPin.blockID, inputPin.io.ioNum, outputPin.blockID, outputPin.io.ioNum)
         this.createConnectionTrace(outputPin, inputPin, inverted)
     }
+    disconnect(inputPin: FunctionBlockPinView<Input>) {
+        this.circuit.disconnectFunctionBlockInput(inputPin.blockID, inputPin.io.ioNum)
+        this.deleteConnectionTrace(inputPin.id)
+    }
 
     createConnectionTrace(outputPin: FunctionBlockPinView<Output>, inputPin: FunctionBlockPinView<Input>, inverted = false) {
-        const trace = new CircuitTrace(this.traceLayer, outputPin, inputPin)
         if (inputPin.io.connection) {
             this.deleteConnectionTrace(inputPin.id)
         }
+        const trace = new CircuitTrace(this.traceLayer, outputPin, inputPin)
         this.traces.set(inputPin.id, trace)
     }
     
     deleteConnectionTrace(id: ID) {
         const trace = this.traces.get(id)
-        console.log('Delete connection', id, trace)
+        if (!trace) return
+        console.log('Delete trace', id, trace)
         trace.delete()
         this.traces.delete(id)
     }
@@ -203,7 +215,6 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
         }
         return outputPin
     }
-
 
 
     // Element info to debug string
@@ -227,7 +238,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
             case 'inputPin':
             case 'outputPin':
                 this.selectedElementsInitPos.set(elem, elem.pos.copy())
-                this.traceLayer.addTrace(this.connectingTrace, elem.absPos, elem.absPos, this.style.colorPinHover)
+                this.traceLayer.addTrace(this.connectingTraceID, elem.absPos, elem.absPos, this.style.colorPinHover)
                 break
         }
     }
@@ -242,13 +253,13 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
                 break
             }
             case 'inputPin': {
-                const absPos = Vec2.div(this.pointer.relativeDownPos, this.scale).add(offset)
-                this.traceLayer.updateTrace(this.connectingTrace, absPos, elem.absPos)
+                const absPos = Vec2.div(this.pointer.relativeDownPos, this.scale).add(offset).sub(vec2(0.5))
+                this.traceLayer.updateTrace(this.connectingTraceID, absPos, elem.absPos)
                 break
             }
             case 'outputPin': {
-                const absPos = Vec2.div(this.pointer.relativeDownPos, this.scale).add(offset)
-                this.traceLayer.updateTrace(this.connectingTrace, elem.absPos, absPos)
+                const absPos = Vec2.div(this.pointer.relativeDownPos, this.scale).add(offset).sub(vec2(0.5))
+                this.traceLayer.updateTrace(this.connectingTraceID, elem.absPos, absPos)
                 break
             }
         }
@@ -270,7 +281,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
                     const inputPin = elem as FunctionBlockPinView<Input>
                     this.connect(outputPin, inputPin)
                 }
-                this.traceLayer.deleteTrace(this.connectingTrace)
+                this.traceLayer.deleteTrace(this.connectingTraceID)
                 this.unselectAll()
                 break
             }
@@ -281,7 +292,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
                     const inputPin = targetElem as FunctionBlockPinView<Input>
                     this.connect(outputPin, inputPin)
                 }
-                this.traceLayer.deleteTrace(this.connectingTrace)
+                this.traceLayer.deleteTrace(this.connectingTraceID)
                 this.unselectAll()
                 break
             }
@@ -349,15 +360,17 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
             this.DOMElement.style.cursor = 'grab'
         }
         // Start selection box
-        else if (this.pointer.targetElem == this.blockArea && ev.buttons == MouseButton.LEFT) { 
+        else if (this.pointer.targetElem == this.blockArea && ev.buttons == MouseButton.LEFT) {
             this.draggingMode = DraggingMode.SELECTION_BOX
+            this.selectionBoxInitPos = this.blockArea.relativePixelPos(this.pointer.downPos)
             this.selectionBox = HTML.domElement(this.blockArea.DOMElement, 'div', {
                 position: 'absolute',
                 backgroundColor: 'rgba(128,128,255,0.2)',
                 border: 'thin solid #88F',
                 pointerEvents: 'none',
-                ...getPositiveRectAttributes(this.pointer.relativeDownPos, this.pointer.dragOffset)
+                ...getPositiveRectAttributes(this.selectionBoxInitPos, this.pointer.dragOffset)
             })
+            console.log('selection start:', this.selectionBoxInitPos)
         }
         // Start dragging selection
         else if (this.pointer.isDragging && this.pointer.downTargetElem?.isDraggable) {
@@ -378,7 +391,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
                 break
             
             case DraggingMode.SELECTION_BOX:
-                Object.assign(this.selectionBox.style, getPositiveRectAttributes(this.pointer.relativeDownPos, this.pointer.dragOffset))
+                Object.assign(this.selectionBox.style, getPositiveRectAttributes(this.selectionBoxInitPos, this.pointer.dragOffset))
                 break
 
             case DraggingMode.DRAG_ELEMENT:
@@ -402,7 +415,7 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
             case DraggingMode.SELECTION_BOX:
                 if (!ev.shiftKey) this.unselectAll()
                 this.blocks.forEach(block => {
-                    const pos = Vec2.div(this.pointer.relativeDownPos, this.scale)
+                    const pos = Vec2.div(this.selectionBoxInitPos, this.scale)
                     const size = Vec2.div(this.pointer.dragOffset, this.scale)
                     if (isElementInsideRect(block, pos, size)) {
                         this.selectElement(block)
@@ -424,7 +437,37 @@ export default class CircuitView extends GUIView<CircuitElement, CircuitStyle>
 
         this.draggingMode = DraggingMode.NONE
     }
+
+    deleteElement(elem: CircuitElement) {
+        console.log('Delete element', elem)
+        switch (elem.type)
+        {
+            case 'inputPin':
+            {
+                this.disconnect(elem as FunctionBlockPinView<Input>)
+            }
+        }
+    }
+
+    ////////////////////////////////
+    //      KEYBOARD HANDLING
+    ////////////////////////////////
+    onKeyDown(ev: KeyboardEvent) {
+        switch(ev.code)
+        {
+            case 'Delete':
+            {
+                this.selectedElements.forEach(elem => this.deleteElement(elem))
+                break
+            }
+        }
+    }
+    
+    onKeyUp(ev: KeyboardEvent) {
+        console.log('Key up', ev.code, ev.key)
+    }
 }
+
 
 function getPositiveRectAttributes(pos: Vec2, size: Vec2)
 {
